replace elective_info
